requirements.txt
prettytable==2.5.0
pandas==1.1.5
pyperclip==1.8.0
requests==2.27.1
unicon==20.2
selenium==4.1.2
webdriver-manager==3.5.3
python-dateutil==2.8.2
beautifulsoup4==4.11.1
pytube==12.1.0
pytest==7.1.2
python-dotenv==0.20.0
pyautogui==0.9.53
pyttsx3==2.90
yagmail==0.15.293
docker==6.0.1
flask==2.2.2
